feat: fully flesh out lo_sentinel
lo_sentinel mode now infers if a large object has been created,
destroyed or modified through one of two ways: spying on
`pg_largeobject` or spying on `lo*` and `be_lo*` function calls to set
the sentinel file. I don't expect fully coverage, but it should be "good
enough".

This is done through an `object_access_hook`, which is exceptionally
useful for this type of thing.